+++ tests/test_connectivity.py
@@ -5,138 +5,6 @@ import pyvista as pv
 import pvgridder as pvg
 
 
-# Fixtures for commonly used meshes
-@pytest.fixture
-def anticline_2d():
-    """Fixture for anticline 2D example mesh."""
-    return pvg.examples.load_anticline_2d()
-
-
-@pytest.fixture
-def anticline_3d():
-    """Fixture for anticline 3D example mesh."""
-    return pvg.examples.load_anticline_3d()
-
-
-@pytest.fixture
-def well_2d():
-    """Fixture for well 2D example mesh."""
-    return pvg.examples.load_well_2d()
-
-
-@pytest.fixture
-def well_2d_voronoi():
-    """Fixture for well 2D voronoi example mesh."""
-    return pvg.examples.load_well_2d(voronoi=True)
-
-
-@pytest.fixture
-def well_3d():
-    """Fixture for well 3D example mesh."""
-    return pvg.examples.load_well_3d()
-
-
-@pytest.fixture
-def well_3d_voronoi():
-    """Fixture for well 3D voronoi example mesh."""
-    return pvg.examples.load_well_3d(voronoi=True)
-
-
-@pytest.fixture
-def topographic_terrain():
-    """Fixture for topographic terrain example mesh."""
-    return pvg.examples.load_topographic_terrain()
-
-
-@pytest.fixture
-def simple_2d_grid():
-    """Fixture for a simple 2D quad grid."""
-    return pv.UnstructuredGrid(
-        {
-            pv.CellType.QUAD: np.array(
-                [
-                    [0, 1, 2, 3],
-                    [1, 4, 5, 2],
-                    [4, 6, 7, 5],
-                ]
-            )
-        },
-        np.array(
-            [
-                [0.0, 0.0, 0.0],
-                [1.0, 0.5, 0.0],
-                [1.0, 0.5, 0.0],
-                [0.0, 1.0, 0.0],
-                [2.0, 0.0, 0.0],
-                [2.0, 1.0, 0.0],
-                [3.0, 0.0, 0.0],
-                [3.0, 1.0, 0.0],
-            ]
-        )
-    )
-
-
-@pytest.fixture
-def simple_3d_grid():
-    """Fixture for a simple 3D hexahedral grid."""
-    return pv.UnstructuredGrid(
-        {
-            pv.CellType.HEXAHEDRON: np.array(
-                [
-                    [0, 1, 2, 3, 4, 5, 6, 7],
-                    [1, 8, 9, 2, 5, 10, 11, 6],
-                    [8, 12, 13, 9, 10, 14, 15, 11],
-                ]
-            )
-        },
-        np.array(
-            [
-                [0.0, 0.0, 0.0],
-                [1.0, 0.0, 0.5],
-                [1.0, 1.0, 0.5],
-                [0.0, 1.0, 0.0],
-                [0.0, 0.0, 1.0],
-                [1.0, 0.0, 0.5],
-                [1.0, 1.0, 0.5],
-                [0.0, 1.0, 1.0],
-                [2.0, 0.0, 0.0],
-                [2.0, 1.0, 0.0],
-                [2.0, 0.0, 1.0],
-                [2.0, 1.0, 1.0],
-                [3.0, 0.0, 0.0],
-                [3.0, 1.0, 0.0],
-                [3.0, 0.0, 1.0],
-                [3.0, 1.0, 1.0],
-            ]
-        )
-    )
-
-
-@pytest.fixture
-def small_quad_grid():
-    """Fixture for a small quad grid with two cells."""
-    return pv.UnstructuredGrid(
-        {
-            pv.CellType.QUAD: np.array(
-                [
-                    [0, 1, 2, 3],
-                    [1, 4, 5, 2],
-                ]
-            )
-        },
-        np.array(
-            [
-                [0.0, 0.0, 0.0],
-                [1.0, 0.0, 0.0],
-                [1.0, 1.0, 0.0],
-                [0.0, 1.0, 0.0],
-                [2.0, 0.0, 0.0],
-                [2.0, 1.0, 0.0],
-            ]
-        )
-    )
-
-
 @pytest.mark.parametrize(
     "mesh_fixture",
     [
