+++ tests/test_properties.py
@@ -6,50 +6,6 @@ import pyvista as pv
 import pvgridder as pvg
 
 
-# Test fixtures
-@pytest.fixture
-def simple_unstructured_grid():
-    """Create a simple unstructured grid with hexahedron and quad cells."""
-    return pv.examples.cells.Tetrahedron() + pv.examples.cells.Hexahedron()
-
-
-@pytest.fixture
-def tetrahedron_grid():
-    """Create an unstructured grid with a tetrahedron cell."""
-    return pv.examples.cells.Tetrahedron()
-
-
-@pytest.fixture
-def pyramid_grid():
-    """Create an unstructured grid with a pyramid cell."""
-    return pv.examples.cells.Pyramid()
-
-
-@pytest.fixture
-def explicit_structured_grid():
-    """Create a simple explicit structured grid."""
-    return pv.ExplicitStructuredGrid(pv.RectilinearGrid([0.0, 1.0, 2.0], [0.0, 1.0, 2.0], [0.0, 1.0, 2.0]))
-
-
-@pytest.fixture
-def structured_grid_1d():
-    """Create a 1D structured grid."""
-    return pv.RectilinearGrid([0.0, 1.0, 2.0], [0.0], [0.0]).cast_to_structured_grid()
-
-
-@pytest.fixture
-def structured_grid_2d():
-    """Create a 2D structured grid."""
-    return pv.RectilinearGrid([0.0, 1.0, 2.0], [0.0, 1.0, 2.0], [0.0]).cast_to_structured_grid()
-
-
-@pytest.fixture
-def structured_grid_3d():
-    """Create a 3D structured grid."""
-    return pv.RectilinearGrid([0.0, 1.0, 2.0], [0.0, 1.0, 2.0], [0.0, 1.0, 2.0]).cast_to_structured_grid()
-
-
-# Test functions
 @pytest.mark.parametrize(
     "mesh_fixture, flatten",
     [
@@ -64,20 +20,16 @@ def structured_grid_3d():
         pytest.param("pyramid_grid", True, id="pyramid_flat"),  # 5 vertices + 1 for cell size
 
         # Example mesh - complex unstructured grid
-        pytest.param(pvg.examples.load_well_2d, False, id="well_2d_nested"),
-        pytest.param(pvg.examples.load_well_2d, True, id="well_2d_flat"),
-        pytest.param(lambda: pvg.examples.load_well_3d(voronoi=True), False, id="voronoi_well_3d"),
-        pytest.param(lambda: pvg.examples.load_well_3d(voronoi=True), True, id="voronoi_well_3d"),
+        pytest.param("well_2d", False, id="well_2d_nested"),
+        pytest.param("well_2d", True, id="well_2d_flat"),
+        pytest.param("well_3d_voronoi", False, id="voronoi_well_3d_nested"),
+        pytest.param("well_3d_voronoi", True, id="voronoi_well_3d_flat"),
     ],
 )
 def test_get_cell_connectivity(request, mesh_fixture, flatten):
     """Test retrieving cell connectivity with different meshes and flatten options."""
     # Get the actual mesh
-    if callable(mesh_fixture):
-        actual_mesh = mesh_fixture()
-
-    else:
-        actual_mesh = request.getfixturevalue(mesh_fixture)
+    actual_mesh = request.getfixturevalue(mesh_fixture)
     
     # Get cell connectivity
     result = pvg.get_cell_connectivity(actual_mesh, flatten=flatten)
@@ -118,8 +70,8 @@ def test_get_cell_connectivity(request, mesh_fixture, flatten):
         pytest.param(pv.examples.cells.Vertex, 0, id="vertex_ugrid"),
         
         # Example meshes
-        pytest.param(pvg.examples.load_anticline_2d, 2, id="anticline_2d"),
-        pytest.param(pvg.examples.load_anticline_3d, 3, id="anticline_3d"),
+        pytest.param("anticline_2d", 2, id="anticline_2d"),
+        pytest.param("anticline_3d", 3, id="anticline_3d"),
     ],
 )
 def test_get_dimension(request, mesh_fixture, expected_dimension):
@@ -127,7 +79,6 @@ def test_get_dimension(request, mesh_fixture, expected_dimension):
     # Get the actual mesh
     if callable(mesh_fixture):
         actual_mesh = mesh_fixture()
-
     else:
         actual_mesh = request.getfixturevalue(mesh_fixture)
